requirements.txt
annotated-types==0.7.0
bibtexparser==1.4.4
blake3==1.0.8
boto3==1.42.61
botocore==1.42.61
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
emmet-core==0.86.3
fonttools==4.61.1
h5py==3.15.1
idna==3.11
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.9
latexcodec==3.0.1
matplotlib==3.10.8
monty==2026.2.18
mp-api==0.46.0
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
orjson==3.11.7
packaging==26.0
palettable==3.3.3
pandas==3.0.1
phonopy==3.0.1
pillow==12.1.1
plotly==6.6.0
pybtex==0.25.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymatgen==2025.10.7
pymatgen-io-validation==0.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.19.1
s3transfer==0.16.0
scipy==1.17.1
six==1.17.0
spglib==2.7.0
symfc==1.6.1
sympy==1.14.0
tabulate==0.10.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uncertainties==3.2.3
urllib3==2.6.3
